+++ proc-macro/src/lib.rs
@@ -1,11 +1,6 @@
 use proc_macro::TokenStream;
-use proc_macro2::TokenStream as TokenStream2;
-use quote::quote;
-use syn::{parse_macro_input, punctuated::Punctuated, ItemFn, ReturnType, Signature, Token, Type};
 
 mod entry;
-use entry::Arg;
-
 /**
 Configures your `main` function to be exported properly.
 
@@ -48,95 +43,5 @@ fn main() {
 */
 #[proc_macro_attribute]
 pub fn entry(attr: TokenStream, input: TokenStream) -> TokenStream {
-    let ItemFn {
-        attrs,
-        vis,
-        sig,
-        block,
-    } = parse_macro_input!(input as ItemFn);
-    if let Err(e) = verify_signature(&sig) {
-        return e.into();
-    };
-
-    let resources = {
-        if attr.is_empty() {
-            None
-        } else {
-            Some(parse_macro_input!(attr with Punctuated::<Arg, Token![,]>::parse_separated_nonempty)
-                .into_iter())
-        }
-    }.map(|args| {
-        quote! {
-            let teensy4_bsp::board::Resources {
-                #(#args),* , ..
-            } = teensy4_bsp::board::t41(teensy4_bsp::board::instances());
-        }
-    });
-
-    quote! {
-        #[no_mangle]
-        #(#attrs)*
-        #vis #sig {
-            #resources
-
-            #block
-        }
-    }
-    .into()
-}
-
-fn verify_signature(signature: &Signature) -> Result<(), TokenStream2> {
-    macro_rules! is {
-        (@some $( $prop: ident ).+, $value: expr) => {
-            is!(@eq $( $prop ).+, Some($value))
-        };
-        (@none $( $prop: ident ).+) => {
-            is!(@eq $( $prop ).+, None)
-        };
-        (@eq $( $prop: ident ).+, $value: expr) => {
-            is!( signature.$( $prop ).+ != $value, &signature.$( $prop ).+ )
-        };
-        (@ne $( $prop: ident ).+, $value: expr) => {
-            is!( signature.$( $prop ).+ == $value, &signature.$( $prop ).+ );
-        };
-        (@empty $( $prop: ident ).+) => {
-            is!( !signature.$( $prop ).+.is_empty(), &signature.$( $prop ).+ );
-        };
-        (@matches $lhs: pat, $( $prop: ident ).+, $out: expr) => {
-            if let $lhs = $( $prop ).+ {
-                is!(@err $out);
-            }
-        };
-        ($bool: expr, $out: expr) => {
-            if $bool {
-                is!(@err $out);
-            }
-        };
-        (@err $out: expr) => {
-            return Err(syn::Error::new_spanned(
-                $out,
-                "this attribute may only be applied to a function with the signature `fn() main -> !`",
-            ).to_compile_error());
-        };
-    }
-
-    is!(@none constness);
-    is!(@none asyncness);
-    is!(@none unsafety);
-    is!(@none abi);
-    is!(@eq ident, "main");
-    is!(@none generics.lt_token);
-    is!(@none generics.gt_token);
-    is!(@empty generics.params);
-    is!(@none generics.where_clause);
-    is!(@empty inputs);
-    is!(@none variadic);
-
-    if let ReturnType::Type(_, boxed) = &signature.output {
-        if let Type::Never(_) = **boxed {
-            return Ok(());
-        }
-    }
-
-    is!(@err &signature.output);
+    entry::entry(attr, input)
 }